refactor(models): extract engine-agnostic ONNX scoring into _onnx_base
The lm-eval scoring logic in the onnxruntime-genai backend (token
windowing, greedy-match loglikelihood, rolling perplexity, stable
log-softmax, tokenizer/genai_config resolution) is independent of the
inference engine driving it. Move it verbatim into _ONNXLMBase so a
second backend can reuse it instead of reimplementing scoring and
drifting numerically.

onnxruntime_genai.py keeps only its engine primitives: _import_genai,
_select_ep, _load, _forward_logits and _generate. _log_softmax stays
re-exported so existing imports keep working. winml.py is untouched and
still inherits everything through ONNXRuntimeGenAILM.

The execution_provider=None default moves to _ONNXLMBase.__init__ with
it, so the genai_config.json providers keep surviving by default as of
86555286 -- the fix belongs to the shared signature now, not to one
backend.

Also pin max_length to the prompt length in _forward_logits. GenAI
otherwise sizes its KV cache from the model's full context_length
(32768 for Qwen1.5), which cost ~34x on single-pass scoring. Logits are
bit-identical; this only drops the surplus allocation.
